organize file to types folder, WIP: setup criteria object for scalable event logic
