+++ api/src/modules/events/pipes/event-status-validation.pipe.ts
@@ -1,5 +1,5 @@
 import { PipeTransform, BadRequestException } from '@nestjs/common';
-import { EventStatus } from '../event-status.enum';
+import { EventStatus } from '../types/event-status.enum';
 
 export class EventStatusValidationPipe implements PipeTransform {
   readonly allowedStatuses = [
