+++ api/src/modules/events/types/criteria.interface.ts
@@ -1,0 +1,8 @@
+import { Location } from './location.interface';
+import { Place } from './place.interface';
+
+export interface Criteria {
+  locations?: Location[];
+
+  places?: Place[];
+}
