+++ api/src/modules/events/types/event-status.enum.ts
