+++ api/src/modules/events/types/location.interface.ts
@@ -1,0 +1,7 @@
+export interface Location {
+  latitude: number;
+
+  longitude: number;
+
+  radius: number;
+}
