+++ api/src/modules/events/types/place.interface.ts
@@ -1,0 +1,5 @@
+export interface Place {
+  keywords: string[];
+
+  radius: number;
+}
